Import packet generation tool
The "pktgen" tool is to be used to generate pcap files which contain traffic
patterns modelled after those which can be found in actual networks. The
generated pcap files can then be used to test the different applications
using them.

The shape of the generated traffic is defined in a declarative way, using
"packet spec files", which look like the following (minimal) example:

    ip { tcp { .flags "", randbytes { .size 50 } } }

The above examples generates IPv4 packets with a TCP payload, and that
in turn with a payload of 50 randomly-generated bytes. The block items
which content with different network encapsulation map to their Scapy
equivalents, and configuration attributes (e.g. ".flags") allowed are
the same accepted by the constructors of the corresponding Scapy
classes.

In order to allow reusing spec files, it is possible to include a spec
file from another using the "spec-file" element. For example the following
reuses the spec file for TCP packets to create spec files for encapsulating
them with IPv4 and IPv6:

    # tcp-randbytes.pktinc
    tcp { .flags "", randbytes { .size 50 } }

    # ipv4-tcp-randbytes.pktspec
    ipv4 { spec-file { .name "tcp-randbytes" } }

    # ipv6-tcp-randbytes.pktspec
    ipv6 { spec-file { .name "tcp-randbytes" } }

For a better use case, check the "imix*" files in the "examples/"
subdirectory.