Script to run a benchmark multiple times and gather stats
The benchrun.lua accepts the number of times to execute a command, and the
command itself. The command should output lines in the format:

  Rate: X MPPS

where X is a positive number (either integer or floating point). The following
calculations are made:

* Minimum rate.
* Maximum rate.
* Average rate.
* Standard error of rate.

If the command outputs more than one "Rate: X MPPS" line, the statistis are
done separately for each of the observed values (average of the first values,
average of the second, and so on), plus an additional report with the same
statistics calculated over the sums of all the observed values per each run
of the command.

Example usage:

  sudo ../tests/benchrun.lua 10 'snabb snsh
    apps/lwaftr/benchmark.lua ../tests/apps/lwaftr/data/binding.table \
    ../tests/apps/lwaftr/data/icmp_on_fail.conf \
    ../tests/apps/lwaftr/data/tcp-frominet-bound-550.pcap \
    ../tests/apps/lwaftr/data/empty.pcap \
    0000:04:00.0 0000:04:00.1'

Notice how the command is a single argument to the script.