benchrun.lua: Close popen()'d file descriptor
Closing file-like objects obtained with io.popen() is needed to ensure that
the child process running at the other side of the pipe finishes its execution
and the resources used by it are freed. Explicitly closing the file-like
object avoid hitting the limit of opened file descriptors when the number
of execution rounds is high and the GC does not kick in often enough.