Support TAP and raw socket devices in "snabb lwaftr run"
This introduces configuration of NICs using a "device kind" selection
mechanism.

On top of the existing options (Intel 82599 and VirtioNet), either side of
the AFTR may use either a RawSocket (using the "raw" device kind) or attach
to an existing TUN/TAP device using the Tap application (using the "tap"
device kind). Note that the TAP devices have to be configured manually using
the "ip" command.

The default is still to use Intel 82500 NICs by passing their PCI addresses
to the --v4/--v4 command line flags. This implicitly selects the "intel10g"
device kind (which, for consistency, can be specified explicitly as well).
The existing --v4-pci/--v6-pci flags are kept for backwards compatibility,
making the following command lines equivalent:

   snabb lwaftr run --v4-pci 0:10:1 --v6-pci 0:10:2 ...
   snabb lwaftr run --v4 0:10:1 --v6 0:10:2 ...
   snabb lwaftr run --v4 intel10g:0:10:1 --v6 intel10g:0:10:2 ...

Similarly, to use VirtioNet devices, the "virtio" device kind can be used,
and the --virtio flag is kep for compatibility. The following command lines
are equivalent:

   snabb lwaftr run --v4-pci 0:10:1 --v6-pci 0:10:2 --virtio ...
   snabb lwaftr run --v4 0:10:1 --v6 0:10:2 --virtio ...
   snabb lwaftr run --v4 virtio:0:10:1 --v6 virtio:0:10:2 ...

Using any of the deprecated command line flags will print a warning to the
standard error output.

Note that with this patch applied it is possible to mix and match devices,
thus allowing to use a different device kinds each endpoint of of the AFTR.
For example, the following uses an Intel 82599 NIC on the Internet side, and
a TAP device on the B4 side of the AFTR (i.e. to provide connectivity to
virtual machines which share a bridge with the TAP device):

    snabb lwaftr run --v4 intel10g:0:10:1 --v6 tap:aftr ...