Support TAP devices in "snabb-lwaftr run-nohw"
Either side of the AFTR may use either a RawSocket (by prefixing the device
name with "raw:") or attach to an existing TUN/TAP device using the Tap
application (by prefixing the device name with "tap:"). Note that the TAP
devices have to be configured manually using the "ip" command.