Refactor lwaftr to access packet fields with accessors
Before, packets were mostly accessed via pointers and offsets.  Now we
abstract just a little bit to work in terms of headers and accessors.
Headers are still pointers, though.  The performance is not worse; if
anything a marginal improvement (0.5% or 1% or so).

We also switch to always refer to IPv4 addresses using native
endianness.  This works better with the binding table, is simpler, and
is not slower.

We simplify some of the binding table lookup functions.  This is
important considering that we need to refactor in the future to batch
lookups.

As a functional change, we also change to use a "drop" function instead
of calling packet.free.  Invoking "return drop(pkt)" signifies to the
reader that we are cleaning up after the packet.  Anywhere there's a
function that handles packets and returns without a drop or a transmit
is likely a bug.  Changing to use "return drop(pkt)" identified a few
places we were leaking the packet.