Allow setting number maximum in-flight fragments for reassembly
This adds a "fragment_cache_max" setting to the Reassembler applications (both
for IPv4 and IPv6), which limits the maximum amount of fragments pending to be
reassembled kept in the fragment cache. When the limit is reached, a random
set pending-to-reassemble fragments is evicted from the cache. The setting is
disabled by default.

The purpose of this feature is avoiding an ever-growing fragment cache which
could end up using all the available memory of the machine, and thus avoiding
potential memory exhaustion DoS situations.