Implement utility functions for doubly linked lists
This adds new apps.lwaftr.lwutil.linked_list_*() which can be used to
manipulate circular doubly linked lists of Lua tables. The functions
add the keys "_linked_list_next" and "_linked_list_prev" to the tables
added to the list, so no additional tables are created internally for
housekeeping. The "handle" to a list is a variable with a reference to
the list's first item. Functions which modify the list may return a
changed list "handle". Quick usage example:

  local list = nil  -- Empty list
  list = linked_list_append(list, { message = "Hello" })
  list = linked_list_append(list, { message = "World" })
  while not linked_list_empty(list) do
    local item = linked_list_first(list)
    list = linked_list_remove(list, item)
    print(item.message)
  end