Reassembly: Evict old fragments in constant time
Modifies the Reassembler apps (both IPv4 and IPv6) to keep a linked list of
fragment lists, and adds a "last seen" timestamp to them as well. The linked
list is always kept updated so the fragments most recently seen are always
at the end of the list. Evicting fragments now only needs to traverse the
first elements of the list. There are two situations in which the list of
in-flight fragments is trimmed:

 * Periodically, up 20 fragment lists are scanned, and those which have
   expired a 60s timer evicted.

 * When trying to add a fragment to the cache, if the limit of maximum
   in-flight packets would be exceeded, up to the oldest 50 fragment
   lists are evicted. This does not take into account the timer, because
   in a potential memory exhaustion situation it's better to ensure that
   the Snabb Switch process will continue running.

In the future, it might be desirable to add configurable values for the
expiration timer and the maximum amount of fragment lists scanned for
eviction.